+++ worker/state.go
@@ -7,10 +7,6 @@ type connectionsState struct {
 	Mutex       sync.Mutex
 }
 
-var connections = connectionsState{
-	Connections: make(map[string]*SockConnection),
-}
-
 func (connections *connectionsState) Add(connection *SockConnection) {
 	connections.Mutex.Lock()
 	defer connections.Mutex.Unlock()
@@ -30,10 +26,6 @@ type usersState struct {
 	Mutex sync.Mutex
 }
 
-var users = usersState{
-	Users: make(map[string][]string),
-}
-
 func (users *usersState) Set(user string, connections []string) {
 	users.Mutex.Lock()
 	defer users.Mutex.Unlock()
@@ -44,10 +36,6 @@ func (users *usersState) Set(user string, connections []string) {
 type channelsState struct {
 	Channels map[string][]string
 	Mutex    sync.Mutex
-}
-
-var channels = channelsState{
-	Channels: make(map[string][]string),
 }
 
 func (channels *channelsState) Set(channel string, connections []string) {