+++ api/disconnect_handler.go
@@ -3,12 +3,23 @@ package main
 import (
 	"github.com/Cretezy/dSock/common"
 	"github.com/Cretezy/dSock/common/protos"
+	"github.com/gin-contrib/requestid"
 	"github.com/gin-gonic/gin"
+	"go.uber.org/zap"
 	"google.golang.org/protobuf/proto"
 	"sync"
 )
 
 func disconnectHandler(c *gin.Context) {
+	logger.Debug("Getting disconnect request",
+		zap.String("requestId", requestid.Get(c)),
+		zap.String("id", c.Query("id")),
+		zap.String("user", c.Query("user")),
+		zap.String("session", c.Query("session")),
+		zap.String("channel", c.Query("channel")),
+		zap.String("keepClaims", c.Query("keepClaims")),
+	)
+
 	resolveOptions := common.ResolveOptions{}
 
 	err := c.BindQuery(&resolveOptions)
@@ -93,6 +104,16 @@ func disconnectHandler(c *gin.Context) {
 	}
 
 	workersWaitGroup.Wait()
+
+	logger.Debug("Disconnected",
+		zap.String("requestId", requestid.Get(c)),
+		zap.Strings("workerIds", workerIds),
+		zap.String("id", resolveOptions.Connection),
+		zap.String("user", resolveOptions.User),
+		zap.String("session", resolveOptions.Session),
+		zap.String("channel", resolveOptions.Channel),
+		zap.Bool("keepClaims", keepClaims),
+	)
 
 	c.AbortWithStatusJSON(200, map[string]interface{}{
 		"success": true,