+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 - Improve query binding
-- Add locking for concurrent state (thanks a lot @abdullah-aghayan!)
+- Add locking for concurrent state (thanks a lot [@abdullah-aghayan](https://github.com/abdullah-aghayan)!)
 
 ## v0.3.0 - 2020-05-08