+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added logging to all requests
 - Change unsubscribe type from 2 to 1 (internal type)
 - Fix debug mode never being active
+- Add log request (`log_requests`) option
 
 ## v0.3.0 - 2020-05-08